+++ src/wallet/wallet.cpp
@@ -2720,7 +2720,7 @@ bool CWallet::CreateTransaction(const std::vector<CRecipient> &vecSend,
     // that fee sniping isn't a problem yet, but by implementing a fix now we
     // ensure code won't be written that makes assumptions about nLockTime that
     // preclude a fix later.
-    txNew.nLockTime = chainActive.Height();
+    txNew.nLockTime = static_cast<uint32_t>(chainActive.Height());
 
     // Secondly occasionally randomly pick a nLockTime even further back, so
     // that transactions that are delayed after signing for whatever reason,
@@ -2730,7 +2730,7 @@ bool CWallet::CreateTransaction(const std::vector<CRecipient> &vecSend,
         txNew.nLockTime = std::max(0, (int)txNew.nLockTime - GetRandInt(100));
     }
 
-    assert(txNew.nLockTime <= (unsigned int)chainActive.Height());
+    assert(txNew.nLockTime <= static_cast<uint32_t>(chainActive.Height()));
     assert(txNew.nLockTime < LOCKTIME_THRESHOLD);
 
     {
